LineChatBot/requirements.txt
click==6.6
Flask==0.11.1
future==0.16.0
gunicorn==19.6.0
itsdangerous==0.24
Janome==0.2.8
Jinja2==2.8
line-bot-sdk==1.0.2
MarkupSafe==0.23
mod-pywebsocket==0.7.9
requests==2.11.1
simplejson==3.10.0
Werkzeug==0.11.11
